* 'master' of git://github.com/ieb/open-experiments: (220 commits)
  KERN-521 fixed bundle startup for the rules repository.
  KERN-737 - removed uploading the jar file
  KERN-736 : CreateTree now allows arrays of integers, string, ..
  KERN-674 : Changed QDoc on chat to annotations.
  KERN-675 : Changed QDoc to felix annotations.
  KERN-521 Moved all the CND files into the bundle startup to remove runtime dependency on Jackrabbit core, which is not scoped to test.
  KERN-521 Fixed and cleaned up warnings in the tests.
  added license headers
  cleaned up pom
  removed deprecated calls
  KERN-615 Added test coverage for the files LinkHandlerTracker
  KERN-615 : Added More unit tests to the files bundle
  KERN-728 Fixing broken build
  KERN-728 fixed build break.
  removed deprecated call
  KERN-728 : Made sure that we're using longs in stead of ints for the search counts.
  Improved doc a bit for the batch service.
  Versioned nodes have a frozenPrimaryType which should be checked as well.
  KERN-727 : Ruby test 577 was failing, now fixed. I added the *@TypeHint=nt:file , this will automatically create the correct node structure (nt:file with a nt:resource child)
  KERN-521 Ported JBoss Drools repository to OSGi and Sling.
  ...

+++ slingtests/osgikernel/bundles/search/src/main/java/org/sakaiproject/nakamura/search/processors/NodeSearchBatchResultProcessor.java
@@ -59,7 +59,7 @@ public class NodeSearchBatchResultProcessor implements
     long total = iterator.getSize();
     long start = SearchUtil.getPaging(request, total);
     
-    int nitems = SearchUtil.intRequestParameter(request,
+    long nitems = SearchUtil.intRequestParameter(request,
         PARAMS_ITEMS_PER_PAGE, DEFAULT_PAGED_ITEMS);
 
     iterator.skip(start);
